backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
click==8.1.8
exceptiongroup==1.3.1
fastapi==0.128.0
h11==0.16.0
idna==3.11
passlib[bcrypt]==1.7.4
pgvector==0.4.1
psycopg==3.2.10
pydantic==2.12.5
pydantic_core==2.41.5
pypdf==6.1.0
python-dotenv==1.1.1
python-jose[cryptography]==3.5.0
python-multipart==0.0.20
requests==2.32.4
SQLAlchemy==2.0.43
starlette==0.49.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.39.0
